ghcn/app/pyproject.toml
[tool.poetry]
name = "climatefind"
version = "0.1.0"
description = ""
authors = ["Justin Spratt <dev.Spratt.Justin@dev.bwater.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.3.1"
pandas = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
